final/web/i_del_lITTLE_daMp/deploy/service/requirements.txt
flask
gunicorn
python-ldap